created the database and inserted values already
