+++ budget_app/assets/adaptive-icon.png
