+++ budget_app/assets/splash.png
