--- budget_app/components/useClientOnlyValue.ts
@@ -1,4 +1,0 @@
-// This function is web-only as native doesn't currently support server (or build-time) rendering.
-export function useClientOnlyValue<S, C>(server: S, client: C): S | C {
-  return client;
-}
