+++ budget_app/package.json
@@ -1,41 +1,22 @@
 {
   "name": "budget_app",
-  "main": "expo-router/entry",
   "version": "1.0.0",
+  "main": "expo/AppEntry.js",
   "scripts": {
     "start": "expo start",
     "android": "expo start --android",
     "ios": "expo start --ios",
-    "web": "expo start --web",
-    "test": "jest --watchAll"
-  },
-  "jest": {
-    "preset": "jest-expo"
+    "web": "expo start --web"
   },
   "dependencies": {
-    "@expo/vector-icons": "^14.0.0",
-    "@react-navigation/native": "^6.0.2",
-    "expo": "~50.0.17",
-    "expo-font": "~11.10.3",
-    "expo-linking": "~6.2.2",
-    "expo-router": "~3.4.10",
-    "expo-splash-screen": "~0.26.5",
-    "expo-status-bar": "~1.11.1",
-    "expo-system-ui": "~2.9.4",
-    "expo-web-browser": "~12.8.2",
+    "expo": "~51.0.7",
+    "expo-status-bar": "~1.12.1",
     "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-native": "0.73.6",
-    "react-native-safe-area-context": "4.8.2",
-    "react-native-screens": "~3.29.0",
-    "react-native-web": "~0.19.6"
+    "react-native": "0.74.1"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
     "@types/react": "~18.2.45",
-    "jest": "^29.2.1",
-    "jest-expo": "~50.0.4",
-    "react-test-renderer": "18.2.0",
     "typescript": "^5.1.3"
   },
   "private": true
